Fix catalog import message loading
