+++ src/features/admin/listings-edit.ts
@@ -16,7 +16,7 @@ import {
 import {
   adminLandingPath,
   CONTENT_MULTIPART,
-  gatedPost,
+  contentMultipartRoute,
   requireContentOr,
   withAuth,
 } from "#routes/auth.ts";
@@ -255,11 +255,6 @@ const parseListingForm = (
   }
   return form;
 };
-
-/** A content-gated multipart POST handler: authenticate, then hand the parsed
- * multipart form to `handle`. Shared by the two multipart admin POSTs (listing
- * create and catalog import). */
-export const contentMultipartRoute = gatedPost(CONTENT_MULTIPART);
 
 /**
  * Handle POST /admin/listing (create listing)
