+++ src/features/auth.ts
@@ -714,6 +714,9 @@ export const gatedPost =
   (request: Request): Promise<Response> =>
     withAuth(request, policy, handle);
 
+/** A content-editor POST whose body may include uploaded files. */
+export const contentMultipartRoute = gatedPost(CONTENT_MULTIPART);
+
 /** A {@link gatedPost} whose handler only needs the parsed form — for POST
  * routes that act on the form alone and never read the session. */
 export const formPost =
