+++ src/ui/templates/admin/listings/form-pages.tsx
@@ -1,6 +1,6 @@
 /* jscpd:ignore-start */
 import { t } from "#i18n";
-import { isBuilderEnabled } from "#routes/admin/builder.ts";
+import { isBuilderEnabled } from "#shared/config.ts";
 import { settings } from "#shared/db/settings.ts";
 import { CsrfForm } from "#shared/forms/csrf-form.tsx";
 import { Flash } from "#shared/forms/flash.tsx";
